requirements/test.txt
-r prod.txt
pytest
pytest-cov
pytest-flask
